Refactor analytics/actions
