+++ components/analytics/actions/correlations/correlations-section.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import { ForestPlot } from "@/components/analytics/actions/forest-plot";
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
+import { ForestPlot } from "@/components/analytics/actions/correlations/forest-plot";
+import { SectionCard } from "@/components/analytics/actions/shared/section-card";
 import type { ActionCorrelations } from "@/lib/analytics/actions-api";
 
 type Props = {
@@ -10,18 +10,17 @@ type Props = {
 
 export function CorrelationsSection({ data }: Props) {
   return (
-    <Card>
-      <CardHeader>
-        <CardTitle className="text-base">What predicts completion?</CardTitle>
-        <p className="text-xs text-muted-foreground">
+    <SectionCard
+      title="What predicts completion?"
+      description={
+        <>
           Each row compares actions that have the factor vs. those that don't. Dot = completion-rate lift, line = 95% CI
           (Newcombe), dot size scales with sample size. Rows need ≥30 in the smaller group to compute a lift. Sample of{" "}
           {data.total_sample.toLocaleString()} tracked actions.
-        </p>
-      </CardHeader>
-      <CardContent>
-        <ForestPlot factors={data.factors} />
-      </CardContent>
-    </Card>
+        </>
+      }
+    >
+      <ForestPlot factors={data.factors} />
+    </SectionCard>
   );
 }
